requirements.txt
langchain==0.3.1
langgraph==0.2.5
flask==3.0.3
google-auth-oauthlib==1.2.0
google-api-python-client==2.126.0
spacy==3.7.2
python-docx==1.1.0
reportlab==4.2.0
requests==2.31.0
python-dotenv==1.0.0
transformers==4.40.0
torch==2.2.0+cpu
huggingface_hub==0.22.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0.tar.gz
langchain_huggingface==0.1.0
pdfplumber==0.11.0